external-api, workers: move endpoint under order_book
