+++ workers/api-server/src/http.rs
@@ -10,9 +10,9 @@ mod task;
 mod wallet;
 
 use admin::{
-    AdminGetAggregateMatchableAmountHandler, AdminGetOrderMatchingPoolHandler,
-    AdminRefreshExternalMatchFeesHandler, AdminRefreshTokenMappingHandler,
-    AdminTriggerSnapshotHandler, AdminWalletMatchableOrderIdsHandler, IsLeaderHandler,
+    AdminGetOrderMatchingPoolHandler, AdminRefreshExternalMatchFeesHandler,
+    AdminRefreshTokenMappingHandler, AdminTriggerSnapshotHandler,
+    AdminWalletMatchableOrderIdsHandler, IsLeaderHandler,
 };
 use async_trait::async_trait;
 use common::types::{
@@ -25,9 +25,8 @@ use external_api::{
     http::{
         admin::{
             ADMIN_ASSIGN_ORDER_ROUTE, ADMIN_CREATE_ORDER_IN_MATCHING_POOL_ROUTE,
-            ADMIN_GET_AGGREGATE_MATCHABLE_AMOUNT_ROUTE, ADMIN_GET_ORDER_MATCHING_POOL_ROUTE,
-            ADMIN_MATCHING_POOL_CREATE_ROUTE, ADMIN_MATCHING_POOL_DESTROY_ROUTE,
-            ADMIN_OPEN_ORDERS_ROUTE, ADMIN_ORDER_METADATA_ROUTE,
+            ADMIN_GET_ORDER_MATCHING_POOL_ROUTE, ADMIN_MATCHING_POOL_CREATE_ROUTE,
+            ADMIN_MATCHING_POOL_DESTROY_ROUTE, ADMIN_OPEN_ORDERS_ROUTE, ADMIN_ORDER_METADATA_ROUTE,
             ADMIN_REFRESH_EXTERNAL_MATCH_FEES_ROUTE, ADMIN_REFRESH_TOKEN_MAPPING_ROUTE,
             ADMIN_TRIGGER_SNAPSHOT_ROUTE, ADMIN_WALLET_MATCHABLE_ORDER_IDS_ROUTE, IS_LEADER_ROUTE,
         },
@@ -37,7 +36,8 @@ use external_api::{
         },
         network::{GET_CLUSTER_INFO_ROUTE, GET_NETWORK_TOPOLOGY_ROUTE, GET_PEER_INFO_ROUTE},
         order_book::{
-            GET_EXTERNAL_MATCH_FEE_ROUTE, GET_NETWORK_ORDERS_ROUTE, GET_NETWORK_ORDER_BY_ID_ROUTE,
+            GET_DEPTH_BY_MINT_ROUTE, GET_EXTERNAL_MATCH_FEE_ROUTE, GET_NETWORK_ORDERS_ROUTE,
+            GET_NETWORK_ORDER_BY_ID_ROUTE,
         },
         price_report::{GET_SUPPORTED_TOKENS_ROUTE, GET_TOKEN_PRICES_ROUTE, PRICE_REPORT_ROUTE},
         task::{GET_TASK_QUEUE_PAUSED_ROUTE, GET_TASK_QUEUE_ROUTE, GET_TASK_STATUS_ROUTE},
@@ -64,7 +64,7 @@ use hyper::{
 };
 use num_bigint::BigUint;
 use num_traits::Num;
-use order_book::GetExternalMatchFeesHandler;
+use order_book::{GetDepthByMintHandler, GetExternalMatchFeesHandler};
 use price_report::{GetSupportedTokensHandler, TokenPricesHandler};
 use rate_limit::WalletTaskRateLimiter;
 use state::State;
@@ -489,6 +489,13 @@ impl HttpServer {
             GetExternalMatchFeesHandler,
         );
 
+        // The "/order_book/depth/:mint" route
+        router.add_admin_authenticated_route(
+            &Method::GET,
+            GET_DEPTH_BY_MINT_ROUTE.to_string(),
+            GetDepthByMintHandler::new(state.clone(), config.clone()),
+        );
+
         // --- Network Routes --- //
 
         // The "/network" route
@@ -547,13 +554,6 @@ impl HttpServer {
             &Method::GET,
             ADMIN_WALLET_MATCHABLE_ORDER_IDS_ROUTE.to_string(),
             AdminWalletMatchableOrderIdsHandler::new(state.clone()),
-        );
-
-        // The "/admin/liquidity/:mint" route
-        router.add_admin_authenticated_route(
-            &Method::GET,
-            ADMIN_GET_AGGREGATE_MATCHABLE_AMOUNT_ROUTE.to_string(),
-            AdminGetAggregateMatchableAmountHandler::new(state.clone(), config.clone()),
         );
 
         // The "/admin/matching_pools/:matching_pool" route
